+++ When/Models/Poll.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DateFinder.Models
+namespace When.Models
 {
     public class Poll
     {
