+++ When/Pages/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-@using DateFinder
-@namespace DateFinder.Pages
+@using When
+@namespace When.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
